requirements.txt
# Runtime
Pillow>=10.0.0

# Build & Distribution (opcional)
pyinstaller>=6.0.0
